+++ src/MinimalApi/Bootstrap/TempSystemsManagerProvider/HackSystemsManagerProcessor.cs
@@ -23,37 +23,6 @@ using Microsoft.Extensions.Configuration;
 
 namespace Amazon.Extensions.Configuration.SystemsManager.Internal
 {
-    // The [current implementation](https://github.com/aws/aws-dotnet-extensions-configuration/blob/aa02b6ee4e4139d6a4f84b1b0742291c8c786c33/src/Amazon.Extensions.Configuration.SystemsManager/Internal/SystemsManagerProcessor.cs#L38-L41) supports JSON values from SecretsManager, but not from SSM. Here we override the default processor and look for SSM param values starting with `"JSON_ENCODED"`. If found, use `JsonConfigurationParser` to 
-    public class HybridParameterProcessor : DefaultParameterProcessor
-    {
-        private const string _jsonSentinel = "JSON_ENCODED";
-
-        public override IDictionary<string, string> ProcessParameters(IEnumerable<Parameter> parameters, string path)
-        {
-            var result = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
-
-            foreach (var parameter in parameters.Where(parameter => IncludeParameter(parameter, path)))
-            {
-                if (parameter.Value.StartsWith(_jsonSentinel))
-                {
-                    var parameterDictionary = JsonConfigurationParser
-                        .Parse(parameter.Value.Substring(_jsonSentinel.Length));
-
-                    foreach (var keyValue in parameterDictionary)
-                    {
-                        result[$"{GetKey(parameter, path)}:{keyValue.Key}"] = keyValue.Value;
-                    }
-                }
-                else
-                {
-                    result[GetKey(parameter, path)] = GetValue(parameter, path);
-                }
-            }
-
-            return result;
-        }
-    }
-
     public class HackSystemsManagerProcessor : ISystemsManagerProcessor
     {
         private const string SecretsManagerPath = "/aws/reference/secretsmanager/";
